Improved USB reliability some more
On the transmit side, if we run out of IRPs, we don't stop; we continue
and simply don't transmit back to the host.
Also, when Reset, we de-initialize not just usb_device.c, we also
de-initialize drv_usb.c.

+++ framework/system/common/sys_module.h
@@ -99,7 +99,7 @@ typedef unsigned short int SYS_MODULE_INDEX;
     type.
 */
 
-typedef int SYS_MODULE_OBJ;
+typedef uintptr_t SYS_MODULE_OBJ;
 
 
 // *****************************************************************************
